Send warning email when fridge left switched off
Send an email warning message when it looks like fridge has been left
switched off. Infer this by analysing the tChamber trend while the
fridge is (supposedly) switched on. Send a follow-up message if/when the
condition clears.

Ditto for the heater (for chambers that have one), except use a
heaterOutput threshold (30%, say) below which it's considered to be OFF.

Also, to allow simplification of code that reads dtStarted, do away with
the idea of setting this gyle parameter to -1. This was used to signify
a beer fridge but ChamberDto has had a dedicated flag for that purpose
for some time. A beer fridge's dtStarted should now be set to some
genuine value or 0.