Adjust exothermic inference logic
Experience suggests that, once fermentation has started, the
brew can become exothermic quite rapidly. Accordingly, adjust
the inference logic from "between 12h and 4 days" to "between
2h and 4 days".

In line with this, change the UI "Gyle Management" form to get
off the fence regarding how "start time" may be used. Recommend
that it be used for "when fermentation started / is expected to
start".