Reduce Status page space for small devices
Status page displays stats as a JSON document. The outer braces
(and the first level of indent) are superfluous. On small devices,
where content often extends beyond the right of the viewport, it's
worth dropping these. For larger screens, however, the extra 'space'
around the content looks good, so keep.