Increase XHR rate limit
Increase burst limit for zone 'default-xhr' from 4 to 10.
Increasing to 5 would be sufficient for the case identified
but best to double that as a safety margin.

The case identified:
Logging on followed immediately by rendering the home page leads
to the XHR request issued by the second gauge widget to exceed
the rate limit. Gauge receives a 503 response and displays zero.

From nginx's error.log:

2023/12/27 09:58:07 [error] 10790#10790:
  *40976 limiting requests, excess: 4.500 by zone "default-xhr",
  client: 192.168.1.1, server: pilot.homebrewer.org,
  request: "GET /tempctrl/guest/chamber/2/beer-temp HTTP/1.1",
  host: "pilot.homebrewer.org",
  referrer: "https://pilot.homebrewer.org/"